+++ web/embed_components.html
@@ -41,7 +41,6 @@
     <div>
       <paper-tabs class="tabs" selected="0" style = "position:relative" class="bottom fit vertical" style="top: inherit">
           <span class="flex"></span>
-          <paper-toggle-button checked></paper-toggle-button>
           <paper-icon-button id = "refresh" icon="refresh"></paper-icon-button>
           <paper-icon-button id = "launch" icon="launch"></paper-icon-button>
       </paper-tabs>
