+++ web/embed_style.html
@@ -28,6 +28,25 @@
     --paper-toggle-button-unchecked-ink-color: var(--text-primary-color);
   }
 
+  #errorToast {
+     bottom: 2px;
+  }
+  html .height-max {
+    height:100%;
+  }
+  paper-toggle-button {
+     padding-top: 10px;
+  }
+  #rightContainer {
+    float:right; 
+    height:100%;
+    position:relative;
+  }
+  #leftContainer {
+    float:left;
+    height:100%;
+    position:relative;
+  }
   .splash {
     background: #333;
     pointer-events: none;
