+++ web/embed_temp.html
@@ -17,7 +17,7 @@
     <script src="packages/dart_pad/bower/webcomponentsjs/webcomponents-lite.min.js"></script>
     <link rel="import" href="imports.html"/>
     <link rel="import" href="polymer/animated-dropdown.html">
-    <link rel="import" href="mobile_style.html"/>
+    <link rel="import" href="embed_style.html"/>
 
     <link href='https://fonts.googleapis.com/css?family=Inconsolata:400,700'
           rel='stylesheet' type='text/css'>
@@ -194,7 +194,7 @@
 
     <script src="packages/codemirror/codemirror.js"></script>
     <script src="packages/browser/dart.js" async></script>
-    <script type="application/dart" src="mobilecopy.dart"></script>
+    <script type="application/dart" src="embed_temp.dart"></script>
     <script async>
       (function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){
             (i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),
